backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "LuizEduardo-service",email = "luizeduardo.service@outlook.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "django (>=5.2.7,<6.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "ruff (>=0.14.1,<0.15.0)",
    "taskipy (>=1.14.1,<2.0.0)"
]


[tool.ruff]
line-length = 79
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.taskipy.tasks]
lint = 'ruff check'
format = 'ruff format'
run = 'python manage.py runserver'
migrate = 'python manage.py migrate'
migration = 'python manage.py makemigrations'
# test = 'pytest -s -x --cov=fastapi_zero_asnyc -vv'
pre_format = 'ruff check --fix'
# pre_test = 'task lint'
# post_test = 'coverage html'